requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
idna==2.10
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.10
